Move editor::Log to its own sources
