+++ facade-lib/CMakeLists.txt
@@ -127,6 +127,7 @@ target_sources(${PROJECT_NAME} PRIVATE
 
   include/facade/editor/common.hpp
   include/facade/editor/inspector.hpp
+  include/facade/editor/log.hpp
 
   src/util/data_provider.cpp
   src/util/geometry.cpp
@@ -169,4 +170,5 @@ target_sources(${PROJECT_NAME} PRIVATE
 
   src/editor/common.cpp
   src/editor/inspector.cpp
+  src/editor/log.cpp
 )
